smartd: Add check of temperature threshold config
When configuring temperature monitoring for hard disks,
if the temperature threshold is not set properly, false
alarms will be generated.
This patch will check the temperature threshold setting
when temperature monitoring is enabled. That is, at the
time of initialization, obtain the temperature threshold
defined by the hard disk manufacturer and compare it with
the setting in the configuration file. If the difference
is large, a prompt message will be printed.